Conflicts:
	mkspecs/linux-icc/qmake.conf
	mkspecs/macx-icc/qmake.conf
	mkspecs/win32-icc/qmake.conf
	src/gui/painting/qgrayraster.c

+++ src/gui/image/qbmphandler_p.h
@@ -121,7 +121,7 @@ private:
     State state;
     BMP_FILEHDR fileHeader;
     BMP_INFOHDR infoHeader;
-    int startpos;
+    qint64 startpos;
 };
 
 QT_END_NAMESPACE
